25.11.9第二次提交

+++ web-ai-code/web-ai-project02/tilas-web-management/src/main/java/com/itheima/service/impl/DeptServiceImpl.java
@@ -1,7 +1,10 @@
 package com.itheima.service.impl;
 
 import com.itheima.mapper.DeptMapper;
+import com.itheima.mapper.EmpMapper;
 import com.itheima.pojo.Dept;
+import com.itheima.pojo.Emp;
+import com.itheima.pojo.EmpQueryParam;
 import com.itheima.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +16,8 @@ import java.util.List;
 public class DeptServiceImpl implements DeptService {
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
 
     @Override
     public List<Dept> findAll() {
@@ -21,7 +26,11 @@ public class DeptServiceImpl implements DeptService {
 
     @Override
     public void deleteById(Integer id) {
-        deptMapper.deleteById(id);
+        if(!empMapper.selectByDeptId(id).isEmpty()){
+            throw new RuntimeException("部门下有员工，不能删除");
+        }else {
+            deptMapper.deleteById(id);
+        }
     }
 
     @Override
